refactor(dashboard): tighten gateway sync and release boundaries
Why this change was needed:
Task 7 moved sync and release orchestration out of API views, but review found the first split introduced awkward function-local imports and an unnecessary gateway/saver module boundary. The final shape keeps gateway sync behavior with GatewayHandler while moving release publishing helpers into the release package.

What changed:
- Moved GatewayData and GatewaySaver into gateway.py and restored GatewayHandler.sync_gateway as the sync entrypoint
- Moved gateway releaser code into biz/release/gateway_releaser.py and added a narrow waiter helper for release polling
- Let release_to_stages and SDK generation own their error conversion instead of duplicating try/except in views
- Updated API callers, import-linter config, and focused tests for the new module boundaries

Problem solved:
The dashboard sync/release workflow now avoids circular-import workarounds, keeps public handler methods stable, and makes the release package own release-specific orchestration.
