+++ src/dashboard/apigateway/apigateway/apis/open/resource_version/views.py
@@ -26,12 +26,10 @@ from apigateway.apis.open.permissions import (
 )
 from apigateway.apps.openapi.models import OpenAPIFileResourceSchemaVersion
 from apigateway.apps.support.models import ResourceDoc, ResourceDocVersion
-from apigateway.biz.gateway import ReleaseError
 from apigateway.biz.release import ReleaseHandler
 from apigateway.biz.resource.importer.openapi import OpenAPIExportManager
 from apigateway.biz.resource_version import ResourceDocVersionHandler, ResourceVersionHandler
 from apigateway.core.models import ResourceVersion, Stage
-from apigateway.utils.exception import LockTimeout
 from apigateway.utils.responses import V1FailJsonResponse, V1OKJsonResponse
 
 from .serializers import (
@@ -124,18 +122,15 @@ class ResourceVersionReleaseApi(generics.CreateAPIView):
         stage_id_to_stage = {stage.id: stage for stage in Stage.objects.filter(id__in=stage_ids)}
         stages = [stage_id_to_stage[stage_id] for stage_id in stage_ids]
         resource_version_obj = ResourceVersion.objects.get(id=data["resource_version_id"])
-        try:
-            ReleaseHandler.release_to_stages(
-                gateway=request.gateway,
-                resource_version=resource_version_obj,
-                stages=stages,
-                username=request.user.username,
-                comment=data["comment"],
-            )
-        except LockTimeout as err:
-            return V1FailJsonResponse(str(err))
-        except ReleaseError as err:
-            return V1FailJsonResponse(str(err))
+        ok, message = ReleaseHandler.release_to_stages(
+            gateway=request.gateway,
+            resource_version=resource_version_obj,
+            stages=stages,
+            username=request.user.username,
+            comment=data["comment"],
+        )
+        if not ok:
+            return V1FailJsonResponse(message)
 
         return V1OKJsonResponse(
             "OK",
