+++ src/dashboard/apigateway/apigateway/biz/gateway/gateway.py
@@ -25,7 +25,7 @@ from typing import Any, Dict, List, Optional
 from django.conf import settings
 from django.db.models import Count
 from django.utils import timezone
-from pydantic import TypeAdapter
+from pydantic import BaseModel, Field, TypeAdapter, field_validator
 
 from apigateway.apps.data_plane.models import DataPlane, GatewayDataPlaneBinding
 from apigateway.apps.metrics.models import StatisticsAppRequestByDay, StatisticsGatewayRequestByDay
@@ -60,6 +60,28 @@ logger = logging.getLogger(__name__)
 # 运营状态查询时间范围，默认 180 天
 # 用于查询网关的运营状态，如果网关在最近 180 天内有请求数据，则认为网关处于活跃状态
 OPERATION_STATUS_DELTA_DAYS = 180
+
+
+class GatewayData(BaseModel):
+    name: str = Field(...)
+    description: str = Field(default="")
+    description_en: Optional[str] = Field(default=None)
+    maintainers: List[str] = Field(default_factory=list)
+    status: int = Field(...)
+    is_public: bool = Field(default=False)
+    gateway_type: Optional[GatewayTypeEnum] = Field(default=None)
+    user_config: Optional[Dict] = Field(default=None)
+    # allow_auth_from_params/allow_delete_sensitive_params 默认值 None，即默认不修改此配置，
+    # 上层如需修改，需明确指定配置值
+    allow_auth_from_params: Optional[bool] = Field(default=None)
+    allow_delete_sensitive_params: Optional[bool] = Field(default=None)
+
+    tenant_mode: Optional[str] = Field(default=None)
+    tenant_id: Optional[str] = Field(default=None)
+
+    @field_validator("gateway_type")
+    def validate_gateway_type(cls, v):  # noqa: N805
+        return GatewayTypeEnum(v) if isinstance(v, int) else v
 
 
 class GatewayHandler:
@@ -113,33 +135,10 @@ class GatewayHandler:
         return queryset.filter(id__in=released_gateway_ids)
 
     @staticmethod
-    def sync_gateway(
-        *,
-        gateway: Optional[Gateway],
-        data: dict,
-        bk_app_code: str,
-        username: str,
-        source: Optional[CallSourceTypeEnum],
-        data_plane_ids: Optional[List[int]],
-    ) -> Gateway:
-        from .saver import GatewayData, GatewaySaver  # noqa: PLC0415
-
-        saver = GatewaySaver(
-            id=gateway and gateway.id,
-            data=TypeAdapter(GatewayData).validate_python(data),
-            bk_app_code=bk_app_code,
-            username=username,
-            source=source,
-            data_plane_ids=data_plane_ids,
-        )
-        return saver.save()
-
-    @staticmethod
     def get_gateway_auth_config(gateway_id: int) -> dict:
         """
         获取网关当前的认证配置
         """
-
         try:
             return GatewayAuthContext().get_config(gateway_id)
         except Context.DoesNotExist:
@@ -191,7 +190,8 @@ class GatewayHandler:
             # 多租户版本，只允许从请求头获取认证信息，如果注册方配置 allow_auth_from_params 为 True，则强制设置为 False
             if allow_auth_from_params and settings.ENABLE_MULTI_TENANT_MODE:
                 logger.warning(
-                    "multi-tenant mode, allow_auth_from_params=True is not supported, force set to False, gateway_id=%s",
+                    "multi-tenant mode, allow_auth_from_params=True is not supported, force set to False, "
+                    "gateway_id=%s",
                     gateway_id,
                 )
                 new_config["allow_auth_from_params"] = False
@@ -224,36 +224,44 @@ class GatewayHandler:
         # 用于标识创建网关的来源
         source: Optional[CallSourceTypeEnum] = None,
     ):
-        # 1. save gateway auth_config
         GatewayHandler.save_auth_config(
             gateway.id,
             user_auth_type=user_auth_type,
             user_conf=user_config,
-            api_type=api_type,
             unfiltered_sensitive_keys=unfiltered_sensitive_keys,
+            api_type=api_type,
             allow_auth_from_params=allow_auth_from_params,
             allow_delete_sensitive_params=allow_delete_sensitive_params,
         )
 
-        # 2. save jwt
-
         GatewayJWTHandler.create_jwt(gateway)
-
-        # 3. create default stage
-
         StageHandler().create_default(gateway, created_by=username, source=source)
-
-        # 4. create default alarm-strategy
-
         create_default_alarm_strategy(gateway, created_by=username)
 
-        # 5. create related app
         if related_app_code:
             GatewayRelatedAppHandler.add_related_app(gateway.id, related_app_code)
 
-        # 6. update gateway app binding
         if app_codes_to_binding is not None:
             GatewayAppBindingHandler.update_gateway_app_bindings(gateway, app_codes_to_binding)
+
+    @staticmethod
+    def sync_gateway(
+        gateway: Optional[Gateway],
+        data: dict,
+        bk_app_code: str,
+        username: str,
+        source: Optional[CallSourceTypeEnum],
+        data_plane_ids: Optional[List[int]],
+    ) -> Gateway:
+        saver = GatewaySaver(
+            id=gateway and gateway.id,
+            data=TypeAdapter(GatewayData).validate_python(data),
+            bk_app_code=bk_app_code,
+            username=username,
+            source=source,
+            data_plane_ids=data_plane_ids,
+        )
+        return saver.save()
 
     @staticmethod
     def bind_to_data_planes(gateway: Gateway, data_plane_ids: List[int], username: str = ""):
@@ -448,3 +456,149 @@ class GatewayHandler:
                 operation_statuses[gateway_id]["status"] = GatewayOperationStatusEnum.INACTIVE.value
 
         return operation_statuses
+
+
+class GatewaySaver:
+    """
+    Gateway saver that handles creating/updating gateways and binding to data planes.
+
+    For new gateways:
+    - If data_plane_ids is provided, bind to those data planes
+    - Otherwise, bind to the 'default' data plane
+    """
+
+    def __init__(
+        self,
+        id: Optional[int],
+        data: GatewayData,
+        bk_app_code: str = "",
+        username: str = "",
+        source: Optional[CallSourceTypeEnum] = None,
+        data_plane_ids: Optional[List[int]] = None,
+    ):
+        self.bk_app_code = bk_app_code
+        self.username = username
+
+        self._gateway = self._get_gateway(id)
+        self._gateway_data = data
+        self._source = source
+        self._data_plane_ids = data_plane_ids
+
+    def _get_gateway(self, gateway_id: Optional[int]) -> Optional[Gateway]:
+        if gateway_id:
+            return Gateway.objects.get(id=gateway_id)
+
+        return None
+
+    def save(self) -> Gateway:
+        # 网关为 None，则新建网关；非 None，则更新网关
+        if not self._gateway:
+            self._create_gateway()
+        else:
+            self._update_gateway()
+
+        assert self._gateway
+
+        return self._gateway
+
+    def _create_gateway(self):
+        # 1. save gateway
+        self._gateway = gateway = Gateway(
+            name=self._gateway_data.name,
+            description=self._gateway_data.description,
+            description_en=self._gateway_data.description_en,
+            maintainers=self._gateway_data.maintainers,
+            status=self._gateway_data.status,
+            is_public=self._gateway_data.is_public,
+            tenant_mode=self._gateway_data.tenant_mode,
+            tenant_id=self._gateway_data.tenant_id,
+            created_by=self.username,
+            updated_by=self.username,
+        )
+        gateway.save()
+
+        # 2. save related data
+        GatewayHandler.save_related_data(
+            gateway=gateway,
+            user_auth_type=settings.DEFAULT_USER_AUTH_TYPE,
+            username=self.username,
+            related_app_code=self.bk_app_code,
+            user_config=self._gateway_data.user_config,
+            unfiltered_sensitive_keys=self._get_gateway_unfiltered_sensitive_keys(gateway.name),
+            api_type=self._gateway_data.gateway_type,
+            allow_auth_from_params=self._gateway_data.allow_auth_from_params,
+            allow_delete_sensitive_params=self._gateway_data.allow_delete_sensitive_params,
+            source=self._source,
+        )
+
+        # 3. bind to data plane(s)
+        self._bind_to_data_planes(gateway)
+
+    def _bind_to_data_planes(self, gateway: Gateway):
+        """Bind newly created gateway to data plane(s)"""
+        data_planes_to_bind: List[DataPlane] = []
+
+        # If data_plane_ids provided, bind to those data planes
+        if self._data_plane_ids:
+            data_plane_id_to_obj = {item.id: item for item in DataPlane.objects.filter(id__in=self._data_plane_ids)}
+            found_ids = set(data_plane_id_to_obj.keys())
+            missing_ids = set(self._data_plane_ids) - found_ids
+            if missing_ids:
+                missing_ids_list = sorted(missing_ids)
+                logger.error("Gateway '%s': invalid data_plane_ids=%s", gateway.name, missing_ids_list)
+                raise ValueError(f"invalid data_plane_ids: {missing_ids_list}")
+
+            data_planes_to_bind.extend(data_plane_id_to_obj[data_plane_id] for data_plane_id in self._data_plane_ids)
+        else:
+            default_data_plane = DataPlane.objects.get_default()
+            data_planes_to_bind.append(default_data_plane)
+
+        # Bind to all resolved data planes
+        for data_plane in data_planes_to_bind:
+            GatewayDataPlaneBinding.objects.bind_gateway_to_data_plane(
+                gateway=gateway,
+                data_plane=data_plane,
+                created_by=self.username or "system",
+            )
+            logger.info(
+                "Bound gateway '%s' to data plane '%s'",
+                gateway.name,
+                data_plane.name,
+            )
+
+    def _update_gateway(self):
+        gateway = self._gateway
+
+        # 1. update gateway
+        gateway.description = self._gateway_data.description
+        gateway.description_en = self._gateway_data.description_en
+        # 更新网关时，仅新增网关管理员，不删除，以防止删除已更新的管理员数据
+        gateway.maintainers = sorted(set(self._gateway_data.maintainers + gateway.maintainers))
+        gateway.is_public = self._gateway_data.is_public
+        gateway.updated_by = self.username
+        gateway.save()
+
+        # 2. update auth config
+        GatewayHandler.save_auth_config(
+            gateway.id,
+            user_conf=self._gateway_data.user_config,
+            unfiltered_sensitive_keys=self._get_gateway_unfiltered_sensitive_keys(gateway.name),
+            api_type=self._gateway_data.gateway_type,
+            allow_auth_from_params=self._gateway_data.allow_auth_from_params,
+            allow_delete_sensitive_params=self._gateway_data.allow_delete_sensitive_params,
+        )
+
+    def _get_gateway_unfiltered_sensitive_keys(self, gateway_name: str) -> Optional[List[str]]:
+        gateway_auth_configs = getattr(settings, "SPECIAL_GATEWAY_AUTH_CONFIGS", None) or {}
+        return gateway_auth_configs.get(gateway_name, {}).get("unfiltered_sensitive_keys")
+
+
+def sync_gateway(
+    gateway: Optional[Gateway],
+    data: dict,
+    bk_app_code: str,
+    username: str,
+    source: Optional[CallSourceTypeEnum],
+    data_plane_ids: Optional[List[int]],
+) -> Gateway:
+    return GatewayHandler.sync_gateway(gateway, data, bk_app_code, username, source, data_plane_ids)
