+++ src/dashboard/apigateway/apigateway/biz/gateway/related_app.py
@@ -64,7 +64,7 @@ class GatewayRelatedAppHandler:
             GatewayRelatedApp.objects.filter(gateway=gateway, bk_app_code__in=app_codes_to_delete).delete()
 
 
-def sync_related_apps(*, gateway: Gateway, bk_app_codes: List[str], username: str) -> List[str]:
+def sync_related_apps(gateway: Gateway, bk_app_codes: List[str], username: str) -> List[str]:
     related_app_codes = GatewayRelatedAppHandler.get_related_app_codes(gateway.id)
     app_codes_to_add = set(bk_app_codes) - set(related_app_codes)
